add player's volume serialize

+++ src/PlayerDocument.h
@@ -15,7 +15,7 @@
 /// value need to searialize.
 @property (nonatomic) int resumeAtReboot;
 @property (nonatomic) int trackSongsWhenPlayStarted;
-@property (nonatomic) int volume;
+@property (nonatomic) float volume;
 @property (nonatomic) int playOrder; //enum PlayOrder
 @property (nonatomic) int playState;//enum PlayStatus
 @property (nonatomic) int fontHeight;
